Improvement resulting of the testing with a problematic image from Johannes
The image was:

- a post operative CT, so the electrodes are segmented as frame
- FoV in SI direction the bars of each z frame are segmented
individually because they are not linked anymore.
- the image was strangely resampled, with a lot of background volume
added (probably registered to an image with larger extend.)

The solutions:

- filter the objects by their distance from the centroid of all objects
above the soft tissue threshold. they have to be at least at a distance
of 30% of the smallest extend of the image.
- the minimal size of the objects for at least one of the direction is
80% of the smallest extend of the image.
- the aspect ratio filter was relaxed a bit, not sure if needed, but the
other two filters seem to do the job.
- because the dimensions now depend on the image extend, the image is
cropped to the bounding box of non-soft-tissue. This is to exclude any
artificial background added.